- move map to externally accessable page
