+++ frontend/src/components/steps/NamesSummary.tsx
@@ -3,11 +3,14 @@ import useFormWizard from '@/hooks/useFormWizard'
 import useStaticTranslation from '@/hooks/useStaticTranslation'
 import { Box, Button, Checkbox, FormControlLabel, Stack, Typography } from '@mui/material'
 import { MedicineItemType } from 'MedicineTypes'
+import { useRouter } from 'next/router'
 import { ChangeEvent, useState } from 'react'
+
 const NamesSummary = () => {
   const [hasMoreProducts, setHasMoreProducts] = useState(false)
   const { stepTo, formData, updateFormData, submitData } = useFormWizard()
   const { t } = useStaticTranslation()
+  const router = useRouter()
 
   const selectedMedicines = formData?.medicines
 
@@ -16,7 +19,7 @@ const NamesSummary = () => {
       stepTo('details')
     } else {
       submitData('map')
-      stepTo('map')
+      router.push('/map')
     }
   }
 
@@ -34,7 +37,7 @@ const NamesSummary = () => {
   }
 
   return (
-    <Stack gap={2} pb={3} alignItems={'center'} width={'100%'} position={'relative'} justifyContent={'space-between'}>
+    <Stack gap={2} pb={2} alignItems={'center'} width={'100%'} position={'relative'} justifyContent={'space-between'}>
       <Typography variant="h1">{t('names_summary_page_title')}</Typography>
       <Box sx={{ display: 'flex', flex: 1, width: '100%', height: '100%', mt: 3 }}>
         <Stack sx={{ width: '100%', borderRadius: '12px', height: 'fit-content', px: 3, py: 1, boxShadow: '0.5px 1px 4px 2px rgba(0, 0, 0, 0.08)' }}>
